refactor: move certificate helper functions to dedicated file
